// autogpt_platform/frontend/src/app/api/__generated__/models/creditTransactionType.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */

export type CreditTransactionType =
  (typeof CreditTransactionType)[keyof typeof CreditTransactionType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreditTransactionType = {
  TOP_UP: "TOP_UP",
  USAGE: "USAGE",
  GRANT: "GRANT",
  REFUND: "REFUND",
  CARD_CHECK: "CARD_CHECK",
} as const;